Clean up the eval params a bit (removing redundant values). (#37)
 * One fewer opposite colored bishop factor (the last non-fixed value corresponding to a pawn delta of 3 was too noisy).
 * Remove unused mobility bonus/ safe mobility tables for pawn and king.
 * Anchor safe mobility tables to the 4th index (middle of the table) instead of the 5th index.
 * Set the anchored defended/undefended checks adjustment value (0th index) to a value of 0.

Then re-tuned on the same data as the previous params.

Aside from the opposite colored bishop factor, this should have equal expressive power. In practice, the tuned params may be a bit worse, or a bit better. The changes in the eval param struct may change caching behavior (likely for the better since we have less unused data).